+++ Sun Position with Energy Conservation/User.h
@@ -6,11 +6,11 @@ private:
 	Planet* planet = nullptr;
 	SciNumber posXrelative, posYrelative, posZrelative;
 	SciNumber posX, posY, posZ;
-	long longitude, latitude;
+	float longitude, latitude;
 	SciNumber height;
 
 public:
-	User(long longitude, long latitude, double height, Planet* planet);
+	User(float longitude, float latitude, double height, Planet* planet);
 	void calculateRelativePosition();
 	void CalculateGlobalPosition();
 	std::string toString();
